% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{activateTerminal}
\alias{activateTerminal}
\title{Activate Terminal}
\usage{
activateTerminal(id = NULL, show = TRUE)
}
\arguments{
\item{id}{The terminal id. The \code{id} is obtained from
\code{\link{getAllTerminals}()}, \code{\link{getVisibleTerminal}()},
or \code{\link{createTerminal}()}. If NULL, the terminal tab will be
selected but no specific terminal will be chosen.}

\item{show}{If TRUE, bring the terminal to front in RStudio.
terminal buffer.}
}
\description{
Ensure terminal is running and optionally bring to front in RStudio.
}
\note{
The \code{activateTerminal} function was added in version 1.1.233 of RStudio.
}
\examples{
\dontrun{
rstudioapi::activateTerminal('Terminal 1', show=TRUE)
}

}
