% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{applyTheme}
\alias{applyTheme}
\title{Apply an Editor Theme to RStudio}
\usage{
applyTheme(name)
}
\arguments{
\item{name}{The unique name of the theme to apply.}
}
\description{
Applies the specified editor theme to RStudio.
}
