% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{askForPassword}
\alias{askForPassword}
\title{Ask the user for a password interactively}
\usage{
askForPassword(prompt = "Please enter your password", getenv = NULL)
}
\arguments{
\item{prompt}{The prompt to be shown to the user.}

\item{getenv}{An optional getenv. When provided, RStudio will check to see if
an environment variable of the name \verb{RSTUDIOAPI_SECRET_<getenv>} is defined;
if so, that environment variable will be used to supply the password.
If the variable is unset, then (in interactive sessions) the user will
be prompted for a password; otherwise, a warning will be shown.}
}
\description{
Ask the user for a password interactively.
}
\details{
RStudio also sets the global \code{askpass} option to the
\code{rstudioapi::askForPassword} function so that it can be invoked in a
front-end independent manner.
}
\note{
The \code{askForPassword} function was added in version 0.99.853 of
RStudio.
}
\examples{

\dontrun{
rstudioapi::askForPassword("Please enter your password")
}

}
