% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bug-report.R
\name{bugReport}
\alias{bugReport}
\title{File an RStudio Bug Report}
\usage{
bugReport()
}
\description{
A utility function to assist with the filing of an RStudio bug report. This
function will pre-populate a template with information useful in
understanding your reported bug.
}
