% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-tools.R
\name{build-tools}
\alias{build-tools}
\alias{buildToolsCheck}
\alias{buildToolsInstall}
\title{Build Tools}
\usage{
buildToolsCheck()

buildToolsInstall(action)
}
\arguments{
\item{action}{The action (as a string) being taken that will require
installation of build tools.}
}
\description{
Check and install build tools as required.
}
\details{
These functions are intended to be used together -- one should
first check whether build tools are available, and when not,
prompt for installation. For example:\preformatted{compile_model <- function(...) {
  if (!rstudioapi::buildToolsCheck())
    rstudioapi::buildToolsInstall("Model compilation")

# proceed with compilation
}
}

The \code{action} parameter is used to communicate (with a prompt) the operation
being performed that requires build tool installation. Setting it to \code{NULL}
or the empty string will suppress that prompt.
}
\note{
The \code{buildToolsCheck()} and \code{buildToolsInstall()} functions were added
with version 1.2.945 of RStudio.
}
