% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/code.R
\name{callFun}
\alias{callFun}
\title{Call an Rstudio API function}
\usage{
callFun(fname, ...)
}
\arguments{
\item{fname}{name of the Rstudio function to call.
See \code{help(package = "Rstudio")}. For a complete list of functions
in the current API.}

\item{...}{Other arguments passed on to the function}
}
\description{
This function will return an error if Rstudio is not running, or the
function is not available. If you want to fall back to different
behavour, use \code{\link{hasFun}}.
}
\examples{
if (rstudioapi::isAvailable()) {
  rstudioapi::callFun("versionInfo")
}
}

