% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{callRemote}
\alias{callRemote}
\title{Make remote call.}
\usage{
callRemote(call, frame, timeout = 10)
}
\arguments{
\item{call}{An unevaluated function call captured by \code{sys.call()}, representing
the original function invocation to be executed remotely.}

\item{frame}{The environment in which to evaluate the arguments of the \code{call}.
This is typically the parent frame captured by \code{parent.frame()} in the calling function.}

\item{timeout}{Timeout (in seconds) if running as a job (see \code{?isJob}).}
}
\value{
The response from the remote execution of the function call.
}
\description{
This function facilitates making a remote call in an RStudio job by serializing
the call and its arguments, sending them to the remote environment, and waiting
for the response.
}
