% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{clearTerminal}
\alias{clearTerminal}
\title{Clear Terminal Buffer}
\usage{
clearTerminal(id)
}
\arguments{
\item{id}{The terminal id. The \code{id} is obtained from
\code{\link{getAllTerminals}()}, \code{\link{getVisibleTerminal}()},
or \code{\link{createTerminal}()}.'}
}
\description{
Clears the buffer for specified terminal.
}
\note{
The \code{clearTerminal} function was added in version 1.1.227 of RStudio.
}
\examples{
\dontrun{
rstudioapi::clearTerminal("Terminal 1")
}

}
