% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{createTerminal}
\alias{createTerminal}
\title{Create a Terminal}
\usage{
createTerminal(id = NULL)
}
\arguments{
\item{id}{The terminal id. When \code{NULL} or blank, the terminal id
will be chosen by the system.}
}
\value{
The terminal identifier as a character vector (\code{NULL} if
  unable to create the terminal or the given terminal identifier is already
  in use).
}
\description{
Create a new Terminal.
}
\note{
The \code{createTerminal} function was added in version 1.1.233 of RStudio.
}
\examples{
\dontrun{
terminalId <- rstudioapi::createTerminal('My Terminal')
}

}
