% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{dictionaries}
\alias{dictionaries}
\alias{dictionariesPath}
\alias{userDictionariesPath}
\title{Interact with RStudio's Dictionaries}
\usage{
dictionariesPath()

userDictionariesPath()
}
\description{
Interact with the \href{https://hunspell.github.io/}{hunspell} dictionaries
used by RStudio for spell checking.
}
\details{
\code{dictionariesPath()} gives a path to the dictionaries installed and
distributed with RStudio.

\code{userDictionariesPath()} gives the path where users can provide their
own custom \code{hunspell} dictionaries. See:

\url{https://support.posit.co/hc/en-us/articles/200551916-Spelling-Dictionaries-in-the-RStudio-IDE}

for more information.
}
\note{
The \code{dictionariesPath()} and \code{userDictionariesPath()} functions were
introduced with RStudio 1.2.1202.
}
