% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stubs.R
\name{document-save}
\alias{document-save}
\alias{documentSave}
\alias{documentSaveAll}
\title{Save Open Documents}
\usage{
documentSave(id = NULL)

documentSaveAll()
}
\arguments{
\item{id}{The IDs of the document(s) to be saved. When \code{NULL},
this is interpreted as a request to save the current open document.}
}
\description{
Save documents currently open in the RStudio IDE. Note that this
function is used only to save titled documents; untitled documents
cannot be saved through this API.
}
\note{
The \code{documentSave} and \code{documentSaveAll} functions
  were added in version 1.1.{TODO}.
}
