% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-methods.R
\name{document_position}
\alias{document_position}
\alias{is.document_position}
\alias{as.document_position}
\title{Create a Document Position}
\usage{
document_position(row, column)

is.document_position(x)

as.document_position(x)
}
\arguments{
\item{row}{The row (using 1-based indexing).}

\item{column}{The column (using 1-based indexing).}

\item{x}{An object coercable to \code{document_position}.}
}
\description{
Creates a \code{document_position}, which can be used to indicate e.g. the
row + column location of the cursor in a document.
}
\concept{location}
