% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{document_range}
\alias{document_range}
\alias{makeRange}
\title{Create a Range}
\usage{
makeRange(start, end = NULL)
}
\arguments{
\item{start}{A \code{\link{position}} indicating the start of the range.}

\item{end}{A \code{\link{position}} indicating the end of the range.}
}
\value{
A \code{document_range} object, which is simply an \code{R} list with
  fields:

\tabular{ll}{
\code{start:}\tab The start position.\cr
\code{end:}\tab The end position.\cr
}
}
\description{
A \code{document_range} is a pair of \code{\link{document_position}} objects,
with each position indicating the \code{start} and \code{end} of the range,
respectively.
}

