% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files-pane.R
\name{filesPaneNavigate}
\alias{filesPaneNavigate}
\title{Navigate to a Directory in the Files Pane}
\usage{
filesPaneNavigate(path)
}
\arguments{
\item{path}{The filesystem path to be shown.}
}
\description{
Navigate to a directory in the Files pane. The contents of that directory
will be listed and shown in the Files pane.
}
