% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{getActiveTerminalId}
\alias{getActiveTerminalId}
\title{Get Current Terminal}
\usage{
getActiveTerminalId()
}
\value{
The terminal identifier as a character vector (\code{NULL} if no
  terminal is currently open).
}
\description{
Get the identifier for the currently active terminal.
}
\note{
The \code{getActiveTerminalId} function was added in version 1.1.227 of RStudio.
}
