% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{getAllTerminalIds}
\alias{getAllTerminalIds}
\title{Get All Terminal Ids}
\usage{
getAllTerminalIds()
}
\value{
The terminal identifiers as a character vector.
}
\description{
Return a character vector containing all the current terminal identifiers.
}
\note{
The \code{getActiveTerminalId} function was added in version 1.1.227 of RStudio.
}
