% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{getAllTerminals}
\alias{getAllTerminals}
\title{Get All Terminal Ids}
\usage{
getAllTerminals()
}
\value{
The terminal identifiers as a character vector.
}
\description{
Return a character vector containing all the current terminal identifiers.
}
\note{
The \code{getAllTerminals} function was added in version 1.1.233 of RStudio.
}
