% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{getIdentityToken}
\alias{getIdentityToken}
\title{Get the User's Identity Token}
\usage{
getIdentityToken()
}
\value{
A list containing:
\describe{
\item{token}{The identity token string.}
\item{expiry}{The token expiry time as a POSIXct datetime object.}
}
Throws an error if the token cannot be retrieved.
}
\description{
Retrieve the user's OpenID Connect identity token for the current Posit
Workbench session, if any. This token can be used to authenticate with
services that trust Workbench's identity provider.
}
\note{
This function works for any IDE running within a Posit Workbench
session, not just RStudio.
}
