% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{getMode}
\alias{getMode}
\title{Report whether RStudio Desktop or RStudio Server is in use}
\usage{
getMode()
}
\value{
"desktop" for RStudio Desktop installations, and
"server" for RStudio Server / RStudio Workbench installations.
}
\description{
Use \code{getMode()} if you need to differentiate between server
and desktop installations of RStudio.
}
