% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{getThemeInfo}
\alias{getThemeInfo}
\title{Retrieve Themes}
\usage{
getThemeInfo(title, message, url = NULL)
}
\description{
Retrieves a list with themes information. Currently, `editor` as
the theme used under the code editor and `global` as the global theme
applied to the main user interface in RStudio.
}
