% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{getThemes}
\alias{getThemes}
\title{Get Theme List}
\usage{
getThemes()
}
\description{
Retrieves a list of the names of all the editor themes installed for
RStudio.
}
\note{
The \code{getThemes} function was introduced in RStudio 1.2.879.
}
