% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{getVersion}
\alias{getVersion}
\title{Determine the version of RStudio}
\usage{
getVersion()
}
\value{
A \code{"numeric_version"} object, giving the version of RStudio in use.
}
\description{
Use \code{getVersion()} to determine the current version of RStudio.
This can be useful for \R packages which need to gate certain functionality
based on the version of RStudio currently available.
}
