% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{getVisibleTerminal}
\alias{getVisibleTerminal}
\title{Get Visible Terminal}
\usage{
getVisibleTerminal()
}
\value{
Terminal selected in the client, if any.
}
\description{
Get Visible Terminal
}
\note{
The \code{getVisibleTerminal} function was added in version 1.1.227 of RStudio.
}
