% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stubs.R
\name{hasColorConsole}
\alias{hasColorConsole}
\title{Check if console supports ANSI color escapes.}
\usage{
hasColorConsole()
}
\value{
\code{TRUE} if ANSI color escapes are supported; \code{FALSE} otherwise.
}
\description{
Check if the RStudio console supports ANSI color escapes.
}
\note{
The \code{hasColorConsole} function was added in version 1.1.216 of
RStudio.
}
\examples{

\dontrun{
if (rstudioapi::hasColorConsole()) {
  message("RStudio console supports ANSI color sequences.")
}

}


}
