% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{highlightUi}
\alias{highlightUi}
\title{Highlight UI Elements within the RStudio IDE}
\usage{
highlightUi(queries)
}
\arguments{
\item{queries}{A list of "query" objects. Each query should be a list with
entries \code{"query"} and \code{"parent"}. See \strong{Queries} for more details.}
}
\description{
This function can be used to highlight UI elements within the RStudio IDE.
UI elements can be selected using query selectors; most commonly, one should
choose to highlight elements based on their IDs when available.
}
\details{
The tool at:\preformatted{Help -> Diagnostics -> Show DOM Elements
}

can be useful for identifying the classes and IDs assigned to the different
elements within RStudio.
}
\note{
The \code{executeCommand} function was introduced in RStudio 1.3.658.
}
\section{Queries}{


Elements are selected using the same queries as through the web
\code{querySelectorAll()} API. See \url{https://developer.mozilla.org/en-US/docs/Web/API/Document/querySelectorAll}
for more details.

For example, to highlight the Save icon within the Source pane, one might
use:\preformatted{
}
}

