\name{isRstudio}
\alias{checkRstudio}
\alias{isRstudio}
\title{Check if RStudio is running?}
\usage{
isRstudio()

checkRstudio()
}
\value{
\code{isRstudio} a boolean; \code{checkRstudio} an error
message if Rstudio is not running
}
\description{
Check if RStudio is running?
}
\examples{
isRstudio()
\donttest{checkRstudio()}
}

