% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{isTerminalRunning}
\alias{isTerminalRunning}
\title{Is Terminal Running}
\usage{
isTerminalRunning(id)
}
\arguments{
\item{id}{The terminal id. The \code{id} is obtained from
\code{\link{getAllTerminals}()}, \code{\link{getVisibleTerminal}()},
or \code{\link{createTerminal}()}.'}
}
\value{
a boolean
}
\description{
Does a terminal have a process associated with it?
}
\note{
The \code{isTerminalRunning} function was added in version 1.1.233 of RStudio.
}
