% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobAddOutput}
\alias{jobAddOutput}
\title{Add Background Job Output}
\usage{
jobAddOutput(job, output, error = FALSE)
}
\arguments{
\item{job}{The ID of the job that has emitted text.}

\item{output}{The text output emitted by the job.}

\item{error}{Whether the output represents an error.}
}
\description{
Adds text output to a background job.
}
\seealso{
Other jobs:
\code{\link[=jobAdd]{jobAdd()}},
\code{\link[=jobAddProgress]{jobAddProgress()}},
\code{\link[=jobGetState]{jobGetState()}},
\code{\link[=jobList]{jobList()}},
\code{\link[=jobRemove]{jobRemove()}},
\code{\link[=jobRunScript]{jobRunScript()}},
\code{\link[=jobSetProgress]{jobSetProgress()}},
\code{\link[=jobSetState]{jobSetState()}},
\code{\link[=jobSetStatus]{jobSetStatus()}}
}
\concept{jobs}
