% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobAddProgress}
\alias{jobAddProgress}
\title{Add Background Job Progress}
\usage{
jobAddProgress(job, units)
}
\arguments{
\item{job}{The ID of the job to update progress for.}

\item{units}{The integer number of new progress units completed.}
}
\description{
Adds incremental progress units to a background job.
}
\seealso{
Other jobs:
\code{\link[=jobAdd]{jobAdd()}},
\code{\link[=jobAddOutput]{jobAddOutput()}},
\code{\link[=jobGetState]{jobGetState()}},
\code{\link[=jobList]{jobList()}},
\code{\link[=jobRemove]{jobRemove()}},
\code{\link[=jobRunScript]{jobRunScript()}},
\code{\link[=jobSetProgress]{jobSetProgress()}},
\code{\link[=jobSetState]{jobSetState()}},
\code{\link[=jobSetStatus]{jobSetStatus()}}
}
\concept{jobs}
