% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobGetState}
\alias{jobGetState}
\title{Get Background Job State}
\usage{
jobGetState(job)
}
\arguments{
\item{job}{The ID of the job.}
}
\description{
Get Background Job State
}
\seealso{
Other jobs:
\code{\link[=jobAdd]{jobAdd()}},
\code{\link[=jobAddOutput]{jobAddOutput()}},
\code{\link[=jobAddProgress]{jobAddProgress()}},
\code{\link[=jobList]{jobList()}},
\code{\link[=jobRemove]{jobRemove()}},
\code{\link[=jobRunScript]{jobRunScript()}},
\code{\link[=jobSetProgress]{jobSetProgress()}},
\code{\link[=jobSetState]{jobSetState()}},
\code{\link[=jobSetStatus]{jobSetStatus()}}
}
\concept{jobs}
