% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobList}
\alias{jobList}
\title{List Background Jobs}
\usage{
jobList()
}
\description{
List any registered background jobs.
}
\seealso{
Other jobs:
\code{\link[=jobAdd]{jobAdd()}},
\code{\link[=jobAddOutput]{jobAddOutput()}},
\code{\link[=jobAddProgress]{jobAddProgress()}},
\code{\link[=jobGetState]{jobGetState()}},
\code{\link[=jobRemove]{jobRemove()}},
\code{\link[=jobRunScript]{jobRunScript()}},
\code{\link[=jobSetProgress]{jobSetProgress()}},
\code{\link[=jobSetState]{jobSetState()}},
\code{\link[=jobSetStatus]{jobSetStatus()}}
}
\concept{jobs}
