% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobRemove}
\alias{jobRemove}
\title{Remove a Background Job}
\usage{
jobRemove(job)
}
\arguments{
\item{job}{The ID of the job to remove.}
}
\description{
Remove a background job from RStudio's Background Jobs pane.
}
\seealso{
Other jobs:
\code{\link[=jobAdd]{jobAdd()}},
\code{\link[=jobAddOutput]{jobAddOutput()}},
\code{\link[=jobAddProgress]{jobAddProgress()}},
\code{\link[=jobGetState]{jobGetState()}},
\code{\link[=jobList]{jobList()}},
\code{\link[=jobRunScript]{jobRunScript()}},
\code{\link[=jobSetProgress]{jobSetProgress()}},
\code{\link[=jobSetState]{jobSetState()}},
\code{\link[=jobSetStatus]{jobSetStatus()}}
}
\concept{jobs}
