% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobSetProgress}
\alias{jobSetProgress}
\title{Set Background Job Progress}
\usage{
jobSetProgress(job, units)
}
\arguments{
\item{job}{The ID of the job to set progress for.}

\item{units}{The integer number of total units of work completed so far.}
}
\description{
Updates the progress for a background job.
}
\seealso{
Other jobs:
\code{\link[=jobAdd]{jobAdd()}},
\code{\link[=jobAddOutput]{jobAddOutput()}},
\code{\link[=jobAddProgress]{jobAddProgress()}},
\code{\link[=jobGetState]{jobGetState()}},
\code{\link[=jobList]{jobList()}},
\code{\link[=jobRemove]{jobRemove()}},
\code{\link[=jobRunScript]{jobRunScript()}},
\code{\link[=jobSetState]{jobSetState()}},
\code{\link[=jobSetStatus]{jobSetStatus()}}
}
\concept{jobs}
