% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobSetState}
\alias{jobSetState}
\title{Set Background Job State}
\usage{
jobSetState(
  job,
  state = c("idle", "running", "succeeded", "cancelled", "failed")
)
}
\arguments{
\item{job}{The ID of the job on which to change state.}

\item{state}{The new job state.}
}
\description{
Changes the state of a background job.
}
\section{States}{


The following states are supported: \describe{ \item{idle}{The job is
waiting to run.} \item{running}{The job is actively running.}
\item{succeeded}{The job has finished successfully.} \item{cancelled}{The
job was cancelled.} \item{failed}{The job finished but did not succeed.} }
}

\seealso{
Other jobs: \code{\link{jobAddOutput}()},
\code{\link{jobAddProgress}()}, \code{\link{jobAdd}()},
\code{\link{jobRemove}()}, \code{\link{jobRunScript}()},
\code{\link{jobSetProgress}()}, \code{\link{jobSetStatus}()}
}
