% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobSetStatus}
\alias{jobSetStatus}
\title{Set Background Job Status}
\usage{
jobSetStatus(job, status)
}
\arguments{
\item{job}{The ID of the job to update.}

\item{status}{Text describing job's new status.}
}
\description{
Update a background job's informational status text.
}
\seealso{
Other jobs:
\code{\link[=jobAdd]{jobAdd()}},
\code{\link[=jobAddOutput]{jobAddOutput()}},
\code{\link[=jobAddProgress]{jobAddProgress()}},
\code{\link[=jobGetState]{jobGetState()}},
\code{\link[=jobList]{jobList()}},
\code{\link[=jobRemove]{jobRemove()}},
\code{\link[=jobRunScript]{jobRunScript()}},
\code{\link[=jobSetProgress]{jobSetProgress()}},
\code{\link[=jobSetState]{jobSetState()}}
}
\concept{jobs}
