% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{killTerminal}
\alias{killTerminal}
\title{Kill Terminal}
\usage{
killTerminal(id)
}
\arguments{
\item{id}{The terminal id. The \code{id} is obtained from
\code{\link{getAllTerminals}()}, \code{\link{getVisibleTerminal}()},
or \code{\link{createTerminal}()}.'}
}
\description{
Kill processes and close a terminal.
}
\note{
The \code{killTerminal} function was added in version 1.1.234 of RStudio.
}
