% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launcher-functions.R
\name{launcherAvailable}
\alias{launcherAvailable}
\title{Check if Workbench Launcher is Available}
\usage{
launcherAvailable()
}
\description{
Check if the Workbench launcher is available and configured to support
Workbench jobs; that is, jobs normally launched by the user through
the RStudio IDE's user interface.
}
\seealso{
Other job-launcher functionality:
\code{\link[=launcherConfig]{launcherConfig()}},
\code{\link[=launcherContainer]{launcherContainer()}},
\code{\link[=launcherControlJob]{launcherControlJob()}},
\code{\link[=launcherGetInfo]{launcherGetInfo()}},
\code{\link[=launcherGetJob]{launcherGetJob()}},
\code{\link[=launcherGetJobs]{launcherGetJobs()}},
\code{\link[=launcherHostMount]{launcherHostMount()}},
\code{\link[=launcherNfsMount]{launcherNfsMount()}},
\code{\link[=launcherPlacementConstraint]{launcherPlacementConstraint()}},
\code{\link[=launcherResourceLimit]{launcherResourceLimit()}},
\code{\link[=launcherSubmitJob]{launcherSubmitJob()}},
\code{\link[=launcherSubmitR]{launcherSubmitR()}}
}
\concept{job-launcher functionality}
