% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launcher-functions.R
\name{launcherContainer}
\alias{launcherContainer}
\title{Define a Workbench Launcher Container}
\usage{
launcherContainer(image, runAsUserId = NULL, runAsGroupId = NULL)
}
\arguments{
\item{image}{The container image to use.}

\item{runAsUserId}{The user id to run as within the container. Defaults to
the container-specified user.}

\item{runAsGroupId}{The group id to run as within the container. Defaults to
the container-specified group.}
}
\description{
Define a launcher container, suitable for use with the \code{container}
argument to \code{\link[=launcherSubmitJob]{launcherSubmitJob()}}.
}
\seealso{
Other job-launcher functionality: 
\code{\link{launcherAvailable}()},
\code{\link{launcherConfig}()},
\code{\link{launcherControlJob}()},
\code{\link{launcherGetInfo}()},
\code{\link{launcherGetJob}()},
\code{\link{launcherGetJobs}()},
\code{\link{launcherHostMount}()},
\code{\link{launcherNfsMount}()},
\code{\link{launcherPlacementConstraint}()},
\code{\link{launcherResourceLimit}()},
\code{\link{launcherSubmitJob}()},
\code{\link{launcherSubmitR}()}
}
\concept{job-launcher functionality}
