% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launcher-functions.R
\name{launcherControlJob}
\alias{launcherControlJob}
\title{Interact with (Control) a Workbench Job}
\usage{
launcherControlJob(
  jobId,
  operation = c("suspend", "resume", "stop", "kill", "cancel")
)
}
\arguments{
\item{jobId}{The job id.}

\item{operation}{The operation to execute. The operation should be one of
\code{c("suspend", "resume", "stop", "kill", "cancel")}. Note that different
launcher plugins support different subsets of these operations -- consult
your launcher plugin documentation to see which operations are supported.}
}
\description{
Interact with a Workbench job.
}
\seealso{
Other job-launcher functionality:
\code{\link[=launcherAvailable]{launcherAvailable()}},
\code{\link[=launcherConfig]{launcherConfig()}},
\code{\link[=launcherContainer]{launcherContainer()}},
\code{\link[=launcherGetInfo]{launcherGetInfo()}},
\code{\link[=launcherGetJob]{launcherGetJob()}},
\code{\link[=launcherGetJobs]{launcherGetJobs()}},
\code{\link[=launcherHostMount]{launcherHostMount()}},
\code{\link[=launcherNfsMount]{launcherNfsMount()}},
\code{\link[=launcherPlacementConstraint]{launcherPlacementConstraint()}},
\code{\link[=launcherResourceLimit]{launcherResourceLimit()}},
\code{\link[=launcherSubmitJob]{launcherSubmitJob()}},
\code{\link[=launcherSubmitR]{launcherSubmitR()}}
}
\concept{job-launcher functionality}
