% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launcher-functions.R
\name{launcherGetInfo}
\alias{launcherGetInfo}
\title{Retrieve Workbench Launcher Information}
\usage{
launcherGetInfo()
}
\description{
Retrieve information about the Workbench launcher, as well as the different clusters
that the launcher has been configured to use.
}
\seealso{
Other job-launcher functionality:
\code{\link[=launcherAvailable]{launcherAvailable()}},
\code{\link[=launcherConfig]{launcherConfig()}},
\code{\link[=launcherContainer]{launcherContainer()}},
\code{\link[=launcherControlJob]{launcherControlJob()}},
\code{\link[=launcherGetJob]{launcherGetJob()}},
\code{\link[=launcherGetJobs]{launcherGetJobs()}},
\code{\link[=launcherHostMount]{launcherHostMount()}},
\code{\link[=launcherNfsMount]{launcherNfsMount()}},
\code{\link[=launcherPlacementConstraint]{launcherPlacementConstraint()}},
\code{\link[=launcherResourceLimit]{launcherResourceLimit()}},
\code{\link[=launcherSubmitJob]{launcherSubmitJob()}},
\code{\link[=launcherSubmitR]{launcherSubmitR()}}
}
\concept{job-launcher functionality}
