% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launcher-functions.R
\name{launcherGetJob}
\alias{launcherGetJob}
\title{Retrieve Workbench Job Information}
\usage{
launcherGetJob(jobId)
}
\arguments{
\item{jobId}{The id of a Workbench job.}
}
\description{
Retrieve information on a Workbench job with id \code{jobId}.
}
\seealso{
Other job-launcher functionality:
\code{\link[=launcherAvailable]{launcherAvailable()}},
\code{\link[=launcherConfig]{launcherConfig()}},
\code{\link[=launcherContainer]{launcherContainer()}},
\code{\link[=launcherControlJob]{launcherControlJob()}},
\code{\link[=launcherGetInfo]{launcherGetInfo()}},
\code{\link[=launcherGetJobs]{launcherGetJobs()}},
\code{\link[=launcherHostMount]{launcherHostMount()}},
\code{\link[=launcherNfsMount]{launcherNfsMount()}},
\code{\link[=launcherPlacementConstraint]{launcherPlacementConstraint()}},
\code{\link[=launcherResourceLimit]{launcherResourceLimit()}},
\code{\link[=launcherSubmitJob]{launcherSubmitJob()}},
\code{\link[=launcherSubmitR]{launcherSubmitR()}}
}
\concept{job-launcher functionality}
