% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launcher-functions.R
\name{launcherHostMount}
\alias{launcherHostMount}
\title{Define a Workbench Launcher Host Mount}
\usage{
launcherHostMount(path, mountPath, readOnly = TRUE)
}
\arguments{
\item{path}{The host path to be mounted.}

\item{mountPath}{The destination path for the mount in the container.}

\item{readOnly}{Boolean; should the path be mounted read-only?}
}
\description{
Define a launcher host mount, suitable for use with the \code{mounts}
argument to \code{\link[=launcherSubmitJob]{launcherSubmitJob()}}. This can
be used to mount a path from the host into the generated container.
}
\seealso{
Other job submission: \code{\link{launcherConfig}()},
\code{\link{launcherContainer}()}, \code{\link{launcherNfsMount}()},
\code{\link{launcherPlacementConstraint}()},
\code{\link{launcherResourceLimit}()}, \code{\link{launcherSubmitJob}()},
\code{\link{launcherSubmitR}()}
}
