% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launcher-functions.R
\name{launcherNfsMount}
\alias{launcherNfsMount}
\title{Define a Workbench Launcher NFS Mount}
\usage{
launcherNfsMount(host, path, mountPath, readOnly = TRUE)
}
\arguments{
\item{host}{The host name, or IP address, of the NFS server.}

\item{path}{The NFS path to be mounted.}

\item{mountPath}{The destination path for the mount in the container.}

\item{readOnly}{Boolean; should the path be mounted read-only?}
}
\description{
Define a launcher NFS mount, suitable for use with the \code{mounts}
argument to \code{\link[=launcherSubmitJob]{launcherSubmitJob()}}. This can
be used to mount a path from a networked filesystem into a newly generated
container.
}
\seealso{
Other job-launcher functionality:
\code{\link[=launcherAvailable]{launcherAvailable()}},
\code{\link[=launcherConfig]{launcherConfig()}},
\code{\link[=launcherContainer]{launcherContainer()}},
\code{\link[=launcherControlJob]{launcherControlJob()}},
\code{\link[=launcherGetInfo]{launcherGetInfo()}},
\code{\link[=launcherGetJob]{launcherGetJob()}},
\code{\link[=launcherGetJobs]{launcherGetJobs()}},
\code{\link[=launcherHostMount]{launcherHostMount()}},
\code{\link[=launcherPlacementConstraint]{launcherPlacementConstraint()}},
\code{\link[=launcherResourceLimit]{launcherResourceLimit()}},
\code{\link[=launcherSubmitJob]{launcherSubmitJob()}},
\code{\link[=launcherSubmitR]{launcherSubmitR()}}
}
\concept{job-launcher functionality}
