% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launcher-functions.R
\name{launcherResourceLimit}
\alias{launcherResourceLimit}
\title{Define a Workbench Launcher Resource Limit}
\usage{
launcherResourceLimit(type, value)
}
\arguments{
\item{type}{The resource limit type. Must be one of cpuCount, cpuFrequency,
cpuSet, cpuTime, memory, memorySwap. Different launcher plugins may support
different subsets of these resource limit types; please consult the plugin
documentation to learn which limits are supported.}

\item{value}{The formatted value of the requested limit.}
}
\description{
Define a launcher resource limit, suitable for use with the
\code{resourceLimits} argument to
\code{\link[=launcherSubmitJob]{launcherSubmitJob()}}.
}
\seealso{
Other job submission:  \code{\link[=launcherConfig]{launcherConfig()}}, \code{\link[=launcherContainer]{launcherContainer()}},
\code{\link[=launcherHostMount]{launcherHostMount()}}, \code{\link[=launcherNfsMount]{launcherNfsMount()}}, \code{\link[=launcherPlacementConstraint]{launcherPlacementConstraint()}},
\code{\link[=launcherResourceLimit]{launcherResourceLimit()}}, \code{\link[=launcherSubmitR]{launcherSubmitR()}}
}
