% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launcher-functions.R
\name{launcherStatusStream}
\alias{launcherStatusStream}
\title{Stream Job Status-Change Events}
\usage{
launcherStatusStream(jobId = "*", listening)
}
\arguments{
\item{jobId}{The job ID. Use \code{"*"} to stream all job status changes.}

\item{listening}{Boolean; start or stop listening for job status changes.}
}
\description{
Start or stop streaming job status events.
}
