% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launcher-functions.R
\name{launcherStreamOutput}
\alias{launcherStreamOutput}
\title{Start / Stop Streaming Job Output}
\usage{
launcherStreamOutput(jobId, listening)
}
\arguments{
\item{jobId}{The job ID.}

\item{listening}{Boolean; start or stop listening for job output.}
}
\description{
Start or stop streaming job output.
}
