% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stubs.R
\name{persistent-values}
\alias{persistent-values}
\alias{setPersistentValue}
\alias{getPersistentValue}
\title{Persistent Keys and Values}
\usage{
setPersistentValue(name, value)

getPersistentValue(name)
}
\arguments{
\item{name}{Key name}

\item{value}{Key value}
}
\value{
The stored value as a character vector (\code{NULL} if no value
  of the specified name is available).
}
\description{
Store persistent keys and values. Storage is per-project, if there is
no project currently active then a global store is used.
}
\note{
The \code{setPersistentValue} and \code{getPersistentValue} functions
 were added in version 1.1.57 of RStudio.
}
