% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{position}
\alias{makePosition}
\alias{position}
\title{Create a Position}
\usage{
makePosition(row, column)
}
\arguments{
\item{row}{The row (using 1-based indexing).}

\item{column}{The column (using 1-based indexing).}
}
\description{
Creates a \code{position}, which can be used to indicate
e.g. the row + column location of the cursor in a document.
}
\details{
The values \code{-Inf} and \code{Inf} can be used to indicate
the 'first' or 'last' row / column, depending on which document
the position is applied to.
}

