% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stubs.R
\name{previewRd}
\alias{previewRd}
\title{Preview an Rd topic in the Help pane}
\usage{
previewRd(rdFile)
}
\arguments{
\item{rdFile}{The path to an \code{.Rd} file.}
}
\description{
Preview an Rd topic in the Help pane.
}
\note{
The \code{previewRd} function was added in version 0.98.191 of
RStudio.
}
\examples{

\dontrun{
rstudioapi::previewRd("~/MyPackage/man/foo.Rd")
}

}
