% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{previewSql}
\alias{previewSql}
\title{Preview SQL statement}
\usage{
previewSql(conn, statement, ...)
}
\arguments{
\item{conn}{The 'DBI' connection to be used to execute this statement.}

\item{statement}{The SQL stament to execute. Either a path to a
file containing a SQL statement or the SQL statement itself.}

\item{...}{Additional arguments to be used in \code{dbGetQuery()}.

Available from RStudio 1.2.600.}
}
\description{
Makes use of 'DBI' and \code{dbGetQuery()} to execute for proview
a SQL statement for a given 'DBI' connection.
}
