% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{range}
\alias{makeRange}
\alias{range}
\title{Create a Range}
\usage{
makeRange(start, end)
}
\arguments{
\item{start}{A \code{\link{position}} indicating the
start of the range.}

\item{end}{A \code{\link{position}} indicating the
end of the range.}
}
\description{
A \code{range} is a pair of \code{\link{position}} objects,
with each position indicating the \code{start} and \code{end}
of the range, respectively.
}

