% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialogs.R
\name{readPreference}
\alias{readPreference}
\title{Read Preference}
\usage{
readPreference(name, default)
}
\arguments{
\item{name}{The name of the preference.}

\item{default}{The default value to use when the preference is not available.}
}
\description{
Reads a user interface preference, useful to rember preferences across
different r sessions for the same user.
}
\note{
The \code{readPreference} function was added in version 1.1.67 of RStudio.
}
