% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{removeTheme}
\alias{removeTheme}
\title{Remove a custom theme from RStudio.}
\usage{
removeTheme(name)
}
\arguments{
\item{name}{The unique name of the theme to remove.}
}
\description{
Remove a custom theme from RStudio.
}
\note{
The \code{removeTheme} function was introduced in RStudio 1.2.879.
}
