% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stubs.R
\name{replaceRanges}
\alias{replaceRanges}
\title{Replace Text within a Document}
\usage{
replaceRanges(ranges, text, id = NULL)
}
\arguments{
\item{ranges}{A list of \code{range} objects, as created by
\code{\link{makeRange}}.}

\item{text}{A character vector, indicating what text should be
inserted at each aforementioned range. This should either
be length one (in which case, this text is applied to each
range specified); otherwise, it should be the same length
as the \code{ranges} list.}

\item{id}{The document id. When \code{NULL} or blank,
the mutation will apply to the currently open, or last
focused, RStudio document. Use the \code{id} returned
from \code{\link{getActiveDocumentContext}()} to ensure
that the operation is applied on the intended document.}
}
\description{
Use this to change the contents of a document open in RStudio.
}
\seealso{
Other Source Document Methods: \code{\link{getActiveDocumentContext}},
  \code{\link{replaceSelection}}
}

