% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stubs.R
\name{replaceSelection}
\alias{replaceSelection}
\title{Replace Selection within a Document}
\usage{
replaceSelection(text, id = "")
}
\arguments{
\item{text}{A character vector, indicating what
text should be inserted. This will be coerced
into a length one character vector as with
\code{paste(text, collapse = "\\n")}.}

\item{id}{The document id. When \code{NULL} or blank,
the mutation will apply to the currently open, or last
focused, RStudio document. Use the \code{id} returned
from \code{\link{getActiveDocumentContext}()} to ensure
that the operation is applied on the intended document.}
}
\description{
Change the selection in a document.
}
\seealso{
Other Source.Document.Methods: \code{\link{getActiveDocumentContext}},
  \code{\link{replaceRanges}}
}

