% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stubs.R
\name{restartSession}
\alias{restartSession}
\title{Restart the R Session}
\usage{
restartSession(command = "")
}
\arguments{
\item{command}{An \R command (as a string) to be run
after restarting \R.}
}
\description{
Restart the RStudio \R session.
}
\note{
The \code{restartSession} function was added in version 1.1.281
of RStudio.
}
