% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stubs.R
\name{savePlotAsImage}
\alias{savePlotAsImage}
\title{Save active RStudio plot image}
\usage{
savePlotAsImage(
  file,
  format = c("png", "jpeg", "bmp", "tiff", "emf", "svg", "eps"),
  width,
  height
)
}
\arguments{
\item{file}{The target file path.}

\item{format}{The Image format.
Must be one of ("png", "jpeg", "bmp", "tiff", "emf", "svg", or "eps").}

\item{width}{The image width, in pixels.}

\item{height}{The image height, in pixels.}
}
\description{
Save the plot currently displayed in the Plots pane as an image.
}
\note{
The \code{savePlotAsImage} function was introduced in RStudio 1.1.57.
}
