% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selections.R
\name{selections}
\alias{selections}
\alias{selectionGet}
\alias{selectionSet}
\title{Manipulate User Selections in the RStudio IDE}
\usage{
selectionGet(id = NULL)

selectionSet(value = NULL, id = NULL)
}
\arguments{
\item{id}{The document ID. When \code{NULL} (the default), the active, or
most-recently edited, document will be used.}

\item{value}{The text contents to set for the selection.}
}
\description{
These functions allow users of the \code{rstudioapi} package to read and write
the user's current selection within the RStudio IDE.
}
