% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{sendToTerminal}
\alias{sendToTerminal}
\title{Send Text to a Terminal}
\usage{
sendToTerminal(text, id)
}
\arguments{
\item{text}{Character vector containing text to be inserted.}

\item{id}{The terminal id. The \code{id} is obtained from
\code{\link{getAllTerminals}()} or  \code{\link{createTerminal}()}.}
}
\description{
Send text to specified terminal.
}
\note{
The \code{sendToTerminal} function was added in version 1.1.227 of RStudio.
}
\examples{
\dontrun{
rstudioapi::sendToTerminal("ls -l\\n", id = "Terminal 1")
}

}
