% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-api.R
\name{setDocumentContents}
\alias{setDocumentContents}
\title{Set the Contents of a Document}
\usage{
setDocumentContents(text, id = NULL)
}
\arguments{
\item{text}{The text to insert into the document.}

\item{id}{The document id. When \code{NULL} or blank,
the mutation will apply to the currently open, or last
focused, RStudio document. Use the \code{id} returned
from \code{\link{getActiveDocumentContext}()} to ensure
that the operation is applied on the intended document.}
}
\description{
Set the contents of a document, deleting any other text existing in the
document previously.
}

