% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stubs.R
\name{setGhostText}
\alias{setGhostText}
\title{Set ghost text}
\usage{
setGhostText(text)
}
\arguments{
\item{text}{The ghost text to set.}
}
\description{
Set ghost text in the current document. The ghost text will be inserted at
the current cursor position. Ghost text can be inserted into the document
by pressing Tab, and will be automatically dismissed if the user navigates
the cursor away.
}
