% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{showTerminal}
\alias{showTerminal}
\title{Bring Terminal to Front}
\usage{
showTerminal(id = NULL)
}
\arguments{
\item{id}{The terminal id. The \code{id} is obtained from
\code{\link{getAllTerminals}()}, \code{\link{getVisibleTerminal}()},
or \code{\link{createTerminal}()}. If NULL, the terminal tab will be
selected but no specific terminal will be chosen.}
}
\description{
Select terminal tab and display requested terminal.
}
\note{
The \code{showTerminal} function was added in version 1.1.227 of RStudio.
}
\examples{
\dontrun{
rstudioapi::showTerminal("Terminal 1")
}

}
