% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{terminalCreate}
\alias{terminalCreate}
\title{Create a Terminal}
\usage{
terminalCreate(id = NULL, show = TRUE)
}
\arguments{
\item{id}{The desired terminal id. When \code{NULL} or blank,
the terminal id will be chosen by the system.}

\item{show}{If FALSE, terminal won't be brought to front}
}
\value{
The terminal identifier as a character vector (\code{NULL} if
  unable to create the terminal or the given terminal identifier is already
  in use).
}
\description{
Create a new Terminal.
}
\note{
The \code{terminalCreate} function was added in version 1.1.305 of RStudio.
}
\examples{
\dontrun{
terminalId <- rstudioapi::terminalCreate('My Terminal')
}

}
