% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{terminalExecute}
\alias{terminalExecute}
\title{Execute Command}
\usage{
terminalExecute(command, args = character(), workingDir = NULL,
  show = TRUE)
}
\arguments{
\item{command}{System command to be invoked, as a character string.}

\item{args}{Arguments to command}

\item{workingDir}{Working directory for command}

\item{show}{If FALSE, terminal won't be brought to front}
}
\value{
The terminal identifier as a character vector (\code{NULL} if
  unable to create the terminal).
}
\description{
Execute a command, showing results in the terminal pane.
}
\note{
The \code{terminalExecute} function was added in version 1.1.305 of RStudio.
}
