% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{terminalKill}
\alias{terminalKill}
\title{Kill Terminal}
\usage{
terminalKill(id)
}
\arguments{
\item{id}{The terminal id. The \code{id} is obtained from
\code{\link{terminalList}()}, \code{\link{terminalVisible}()},
\code{\link{terminalCreate}()}, or \code{\link{terminalExecute}()}.}
}
\description{
Kill processes and close a terminal.
}
\note{
The \code{terminalKill} function was added in version 1.1.350 of RStudio.
}
