% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{terminalList}
\alias{terminalList}
\title{Get All Terminal Ids}
\usage{
terminalList()
}
\value{
The terminal identifiers as a character vector.
}
\description{
Return a character vector containing all the current terminal identifiers.
}
\note{
The \code{terminalList} function was added in version 1.1.350 of RStudio.
}
