% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{terminalRunning}
\alias{terminalRunning}
\title{Is Terminal Running}
\usage{
terminalRunning(id)
}
\arguments{
\item{id}{The terminal id. The \code{id} is obtained from
\code{\link{terminalList}()}, \code{\link{terminalVisible}()},
or \code{\link{terminalCreate}()}.'}
}
\value{
a boolean
}
\description{
Does a terminal have a process associated with it?
}
\note{
The \code{terminalRunning} function was added in version 1.1.305 of RStudio.
}
