% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{terminalVisible}
\alias{terminalVisible}
\title{Get Visible Terminal}
\usage{
terminalVisible()
}
\value{
Terminal selected in the client, if any.
}
\description{
Get Visible Terminal
}
\note{
The \code{terminalVisible} function was added in version 1.1.305 of RStudio.
}
