% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{unregisterCommandCallback}
\alias{unregisterCommandCallback}
\title{Unregister Command Callback}
\usage{
unregisterCommandCallback(handle)
}
\arguments{
\item{handle}{The registration handle to remove.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Removes a previously registered command callback.
}
\note{
The \code{unregisterCommandCallback} function was introduced in RStudio 1.4.1589.
}
