% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialogs.R
\name{updateDialog}
\alias{updateDialog}
\title{Updates a Dialog Box}
\usage{
updateDialog(...)
}
\arguments{
\item{...}{Named parameters and values to update a dialog box.}
}
\description{
Updates specific properties from the current dialog box.
}
\details{
Currently, the only dialog with support for this action is the
New Connection dialog in which the code preview can be
updated through this API.

\preformatted{
    updateDialog(code = "con <- NULL")
}
}

