% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialogs.R
\name{writePreference}
\alias{writePreference}
\title{Write Preference}
\usage{
writePreference(name, value)
}
\arguments{
\item{name}{The name of the preference.}

\item{value}{The value of the preference.}
}
\description{
Writes a user interface preference, useful to rember preferences across
different r sessions for the same user.
}

