% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{askForPassword}
\alias{askForPassword}
\title{Ask the user for a password interactively}
\usage{
askForPassword(prompt = "Please enter your password:", name = NULL, tag = NULL)
}
\arguments{
\item{prompt}{The prompt to be shown to the user.}

\item{name}{The name to associate with the secret. This name will be used
to locate the requested secret; e.g. when requested via the \code{keyring}
package.}

\item{tag}{An optional tag, used to assist \code{rstudioapi} in reading secrets
in non-interactive \R sessions. Currently, when provided, \code{rstudioapi}
will check if an environment variable of the name \verb{RSTUDIOAPI_SECRET_<KEY>}
is defined; if so, that environment variable will be used to supply the
password. If the variable is unset, then (in interactive sessions) the user
will be prompted for a password; otherwise, a warning will be shown.}
}
\description{
Ask the user for a password interactively. A dialog requesting a password
from the user will be shown, and the value of the retrieved password will
be returned.
}
\details{
RStudio also sets the global \code{askpass} option to the
\code{\link[=askForPassword]{askForPassword()}} function so that it can be invoked in a
front-end independent manner.
}
\note{
The \code{askForPassword} function was added in version 0.99.853 of
RStudio.
}
\examples{

\dontrun{
rstudioapi::askForPassword("Please enter your password:")
}

}
