% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stubs.R
\name{askForPassword}
\alias{askForPassword}
\title{Ask the user for a password interactively}
\usage{
askForPassword(prompt = "Please enter your password")
}
\arguments{
\item{prompt}{The prompt to be shown to the user.}
}
\description{
Ask the user for a password interactively.
}
\details{
RStudio also sets the global \code{askpass} option to the
\code{rstudioapi::askForPassword} function so that it can be invoked in a
front-end independent manner.
}
\note{
The \code{askForPassword} function was added in version 0.99.853 of
RStudio.
}
\examples{

\dontrun{
rstudioapi::askForPassword("Please enter your password")
}

}
