% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialogs.R
\name{askForSecret}
\alias{askForSecret}
\title{Prompt user for secret}
\usage{
askForSecret(
  name,
  message = paste(name, ":", sep = ""),
  title = paste(name, "Secret")
)
}
\arguments{
\item{name}{The name of the secret.}

\item{message}{A character vector with the contents to display in the main
dialog area.}

\item{title}{The title to display in the dialog box.}
}
\description{
Request a secret from the user. If the \code{keyring} package is installed, it
will be used to cache requested secrets.
}
\note{
The \code{askForSecret} function was added in version 1.1.419 of
RStudio.
}
