% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialogs.R
\name{askForSecret}
\alias{askForSecret}
\title{Show Prompt for Secret Dialog}
\usage{
askForSecret(name, message = paste(name, ":", sep = ""), title = paste(name,
  "Secret"))
}
\arguments{
\item{name}{The name of the secret.}

\item{message}{A character vector with the contents to display in
the main dialog area.}

\item{title}{The title to display in the dialog box.}
}
\description{
Shows a dialog box asking for a secret with support to remember such secret
using the 'keyring' package.
}
\note{
The \code{askForSecret} function was added in version 1.1.419 of RStudio.
}
