% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{clearTerminal}
\alias{clearTerminal}
\title{Clear Terminal Buffer}
\usage{
clearTerminal(id = NULL)
}
\arguments{
\item{id}{The terminal id. When \code{NULL} or blank,
the currently open, or last focused, RStudio terminal will be cleared.
Use the \code{id} returned from \code{\link{getActiveTerminalId}()}
or \code{\link{createTerminal}()} to ensure the intended terminal is cleared.}
}
\description{
Clears the buffer for specified terminal.
}
\note{
The \code{clearTerminal} function was added in version 1.1.227 of RStudio.
}
\examples{
\dontrun{
rstudioapi::clearTerminal("Terminal 1")
}

}
