% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-methods.R
\name{document_range}
\alias{as.document_range}
\alias{document_range}
\alias{is.document_range}
\title{Create a Range}
\usage{
document_range(start, end = NULL)

is.document_range(x)

as.document_range(x)
}
\arguments{
\item{start}{A \code{\link{document_position}} indicating the start of the
range.}

\item{end}{A \code{\link{document_position}} indicating the end of the range.}

\item{x}{An object coercable to \code{document_range}.}
}
\value{
An \R \code{list} with class \code{document_range} and fields:

\tabular{ll}{
\code{start:}\tab The start position.\cr
\code{end:}\tab The end position.\cr
}
}
\description{
A \code{document_range} is a pair of \code{\link{document_position}} objects,
with each position indicating the \code{start} and \code{end} of the range,
respectively.
}

