% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stubs.R
\name{getActiveDocumentContext}
\alias{getActiveDocumentContext}
\title{Get the Active Document Context}
\usage{
getActiveDocumentContext()
}
\value{
A \code{data.frame} with elements:
\tabular{ll}{
\code{id:}\tab The document ID.\cr
\code{path:}\tab The path to the document on disk.\cr
\code{contents:}\tab The contents of the document.\cr
\code{selections:}\tab The \code{\link{selection}}s in the document. When the document has multiple selections active, the first selection listed is the primary, or main, selection.\cr
}
}
\description{
Returns information about the currently active
RStudio document.
}
\seealso{
Other Source Document Methods: \code{\link{replaceRanges}},
  \code{\link{replaceSelection}}
}

