% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stubs.R
\name{getActiveDocumentContext}
\alias{getActiveDocumentContext}
\title{Get the Active Document Context}
\usage{
getActiveDocumentContext()
}
\value{
A \code{data.frame} with elements:
\tabular{ll}{
\code{id:}\tab The document ID.\cr
\code{path:}\tab The path to the document on disk.\cr
\code{contents:}\tab The contents of the document.\cr
\code{selection:}\tab A \code{list} of selections. See details for more information.\cr
}
}
\description{
Returns information about the currently active
RStudio document.
}
\details{
The \code{selection} field returns is a list of document selection objects.
A document selection is just a pairing of a document \code{range}, and the
\code{text} within that range.
}
\note{
The \code{getActiveDocumentContext} function was added with version 0.99.796
of RStudio.
}
\seealso{
Other Source Document Methods: \code{\link{insertText}}
}

