% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stubs.R
\name{getActiveProject}
\alias{getActiveProject}
\title{Get Path to Active RStudio Project}
\usage{
getActiveProject()
}
\value{
The path to the current project, or \code{NULL} if no project is
currently open.
}
\description{
Get the path to the active RStudio project (if any).
}
\note{
The \code{getActiveProject} function was added in version 0.99.854 of
RStudio.
}
