% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{getDelegatedAzureToken}
\alias{getDelegatedAzureToken}
\title{OAuth2 Tokens for Delegated Azure Resources}
\usage{
getDelegatedAzureToken(resource)
}
\arguments{
\item{resource}{The name of an Azure resource or service, normally a URL.}
}
\value{
A list containing the OAuth2 token details. Throws an error if unavailable.
}
\description{
When Workbench is using Azure Active Directory for sign-in, this function can
return an OAuth2 token for a service Workbench users have delegated access
to. This requires configuring delegated permissions in Azure itself.
}
\examples{
\dontrun{
getDelegatedAzureToken("https://storage.azure.com")
}
}
