% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{getTerminalBuffer}
\alias{getTerminalBuffer}
\title{Get Terminal Buffer}
\usage{
getTerminalBuffer(id, stripAnsi = TRUE)
}
\arguments{
\item{id}{The terminal id. The \code{id} is obtained from
\code{\link{getAllTerminals}()}, \code{\link{getVisibleTerminal}()},
or \code{\link{createTerminal}()}.'}

\item{stripAnsi}{If FALSE, don't strip out Ansi escape sequences before returning
terminal buffer.}
}
\value{
The terminal contents, one line per row.
}
\description{
Returns contents of a terminal buffer.
}
\note{
The \code{getTerminalBuffer} function was added in version 1.1.234 of RStudio.
}
