% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{getTerminalContext}
\alias{getTerminalContext}
\title{Retrieve Information about RStudio Terminals}
\usage{
getTerminalContext(id)
}
\arguments{
\item{id}{The terminal id. The \code{id}(s) are obtained from
\code{\link{getAllTerminals}()} or  \code{\link{createTerminal}()}.}
}
\value{
A \code{list} with elements:
\tabular{ll}{
\code{handle} \tab the internal handle\cr
\code{caption} \tab caption (terminal identifier)\cr
\code{title} \tab title set by the shell\cr
\code{running} \tab is terminal process executing\cr
\code{busy} \tab is terminal running a program\cr
\code{connection} \tab websockets or rpc\cr
\code{sequence} \tab creation sequence\cr
\code{lines} \tab lines of text in terminal buffer\cr
\code{cols} \tab columns in terminal\cr
\code{rows} \tab rows in terminal\cr
\code{pid} \tab process id of terminal shell\cr
}
}
\description{
Returns information about RStudio terminal instances.
}
\note{
The \code{getTerminalContext} function was added in version 1.1.227 of RStudio.
}
