% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/code.R
\name{getVersion}
\alias{getVersion}
\title{Return the current version of the RStudio API}
\usage{
getVersion()
}
\value{
A \code{\link{numeric_version}} which you can compare to a string
  and get correct results.
}
\description{
Return the current version of the RStudio API
}
\examples{
\dontrun{
if (rstudioapi::getVersion() < "0.98.100") {
  message("Your version of RStudio is quite old")
}
}
}

