% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stubs.R
\name{hasColorConsole}
\alias{hasColorConsole}
\title{Check if Console Supports ANSI Color Escapes}
\usage{
hasColorConsole()
}
\value{
a boolean
}
\description{
Check if Console Supports ANSI Color Escapes
}
\note{
The \code{hasColorConsole} function was added in version 1.1.216
of RStudio.
}
\examples{
\dontrun{
if (rstudioapi::hasColorConsole()) {
  message("RStudio console supports ANSI color sequences.")
}
}

}
