% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/code.R
\name{hasFun}
\alias{findFun}
\alias{hasFun}
\title{Exists/get for Rstudio functions}
\usage{
hasFun(name, version_needed = NULL, ...)

findFun(name, version_needed = NULL, ...)
}
\arguments{
\item{name}{name of object to look for}

\item{version_needed}{An optional version specification. If supplied,
ensures that Rstudio is at least that version. This is useful if
function behaviour has changed over time.}

\item{...}{other arguments passed on to \code{\link[base]{exists}} and
\code{\link[base]{get}}}
}
\description{
These are specialised versions of \code{\link[base]{get}} and
\code{\link[base]{exists}} that look in the rstudio package namespace.
If Rstudio is not running,  \code{exists} will return \code{FALSE},
and \code{getRstudio} will raise an error.
}
\examples{
rstudioapi::hasFun("viewer")
}

