% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/code.R
\name{isAvailable}
\alias{isAvailable}
\alias{verifyAvailable}
\title{Check if RStudio is running.}
\usage{
isAvailable(version_needed = NULL)

verifyAvailable(version_needed = NULL)
}
\arguments{
\item{version_needed}{An optional version specification. If supplied,
ensures that RStudio is at least that version.}
}
\value{
\code{isAvailable} a boolean; \code{verifyAvailable} an error message
  if RStudio is not running
}
\description{
Check if RStudio is running.
}
\examples{
rstudioapi::isAvailable()
\dontrun{rstudioapi::verifyAvailable()}
}

