% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{isJob}
\alias{isJob}
\title{Detect RStudio Jobs}
\usage{
isJob()
}
\value{
Boolean; \code{TRUE} if this is an RStudio job.
}
\description{
Use this function to detect whether RStudio is running an R "job".
These jobs are normally used for actions taken in the Jobs tab, as well
as within the \R build pane.
}
\details{
This function is primarily intended to be used by package authors, who
need to customize the behavior of their methods when run within an
RStudio job.
}
