% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{isTerminalBusy}
\alias{isTerminalBusy}
\title{Is Terminal Busy}
\usage{
isTerminalBusy(id = NULL)
}
\arguments{
\item{id}{The terminal id. When \code{NULL} or blank,
the currently open, or last focused, RStudio terminal will be checked.
Use the \code{id} returned from \code{\link{getActiveTerminalId}()} or
\code{\link{createTerminal}()} to ensure the intended terminal
is queried.}
}
\value{
a boolean
}
\description{
Is a terminal reporting that it is busy?
}
\note{
The \code{isTerminalBusy} function was added in version 1.1.227 of RStudio.
}
