% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{isTerminalBusy}
\alias{isTerminalBusy}
\title{Is Terminal Busy}
\usage{
isTerminalBusy(id)
}
\arguments{
\item{id}{The terminal id. The \code{id} is obtained from
\code{\link{getAllTerminals}()}, \code{\link{getVisibleTerminal}()},
or \code{\link{createTerminal}()}.'}
}
\value{
a boolean
}
\description{
Are terminals reporting that they are busy?
}
\note{
The \code{isTerminalBusy} function was added in version 1.1.233 of RStudio.
}
