% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobAddOutput}
\alias{jobAddOutput}
\title{Add Background Job Output}
\usage{
jobAddOutput(job, output, error = FALSE)
}
\arguments{
\item{job}{The ID of the job that has emitted text.}

\item{output}{The text output emitted by the job.}

\item{error}{Whether the output represents an error.}
}
\description{
Adds text output to a background job.
}
\seealso{
Other jobs: \code{\link{jobAddProgress}()}, \code{\link{jobAdd}()},
\code{\link{jobRemove}()}, \code{\link{jobRunScript}()},
\code{\link{jobSetProgress}()}, \code{\link{jobSetState}()},
\code{\link{jobSetStatus}()}
}
