% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobAddProgress}
\alias{jobAddProgress}
\title{Add Job Progress}
\usage{
jobAddProgress(job, units)
}
\arguments{
\item{job}{The ID of the job to update progress for.}

\item{units}{The integer number of new progress units completed.}
}
\description{
Adds incremental progress units to a job.
}
\seealso{
Other jobs: \code{\link{jobAddOutput}},
  \code{\link{jobAdd}}, \code{\link{jobRemove}},
  \code{\link{jobRunScript}}, \code{\link{jobSetProgress}},
  \code{\link{jobSetState}}, \code{\link{jobSetStatus}}
}
\concept{jobs}
