% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobGetState}
\alias{jobGetState}
\title{Get Background Job State}
\usage{
jobGetState(job)
}
\arguments{
\item{job}{The ID of the job.}
}
\description{
Get Background Job State
}
\seealso{
Other jobs: 
\code{\link{jobAdd}()},
\code{\link{jobAddOutput}()},
\code{\link{jobAddProgress}()},
\code{\link{jobList}()},
\code{\link{jobRemove}()},
\code{\link{jobRunScript}()},
\code{\link{jobSetProgress}()},
\code{\link{jobSetState}()},
\code{\link{jobSetStatus}()}
}
\concept{jobs}
