% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobList}
\alias{jobList}
\title{List Background Jobs}
\usage{
jobList()
}
\description{
List any registered background jobs.
}
\seealso{
Other jobs: 
\code{\link{jobAdd}()},
\code{\link{jobAddOutput}()},
\code{\link{jobAddProgress}()},
\code{\link{jobGetState}()},
\code{\link{jobRemove}()},
\code{\link{jobRunScript}()},
\code{\link{jobSetProgress}()},
\code{\link{jobSetState}()},
\code{\link{jobSetStatus}()}
}
\concept{jobs}
