% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobRemove}
\alias{jobRemove}
\title{Remove a Background Job}
\usage{
jobRemove(job)
}
\arguments{
\item{job}{The ID of the job to remove.}
}
\description{
Remove a background job from RStudio's Background Jobs pane.
}
\seealso{
Other jobs: 
\code{\link{jobAdd}()},
\code{\link{jobAddOutput}()},
\code{\link{jobAddProgress}()},
\code{\link{jobGetState}()},
\code{\link{jobList}()},
\code{\link{jobRunScript}()},
\code{\link{jobSetProgress}()},
\code{\link{jobSetState}()},
\code{\link{jobSetStatus}()}
}
\concept{jobs}
