% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobSetProgress}
\alias{jobSetProgress}
\title{Set Background Job Progress}
\usage{
jobSetProgress(job, units)
}
\arguments{
\item{job}{The ID of the job to set progress for.}

\item{units}{The integer number of total units of work completed so far.}
}
\description{
Updates the progress for a background job.
}
\seealso{
Other jobs: \code{\link{jobAddOutput}()},
\code{\link{jobAddProgress}()}, \code{\link{jobAdd}()},
\code{\link{jobRemove}()}, \code{\link{jobRunScript}()},
\code{\link{jobSetState}()}, \code{\link{jobSetStatus}()}
}
