% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobSetProgress}
\alias{jobSetProgress}
\title{Set Job Progress}
\usage{
jobSetProgress(job, units)
}
\arguments{
\item{job}{The ID of the job to set progress for.}

\item{units}{The integer number of total units of work completed so far.}
}
\description{
Updates the progress for a job.
}
