% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobSetStatus}
\alias{jobSetStatus}
\title{Set Job Status}
\usage{
jobSetStatus(job, status)
}
\arguments{
\item{job}{The ID of the job to update}

\item{status}{Text describing job's new status}
}
\description{
Update a job's informational status text
}
