% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobSetStatus}
\alias{jobSetStatus}
\title{Set Job Status}
\usage{
jobSetStatus(job, status)
}
\arguments{
\item{job}{The ID of the job to update.}

\item{status}{Text describing job's new status.}
}
\description{
Update a job's informational status text.
}
\seealso{
Other jobs: \code{\link{jobAddOutput}},
  \code{\link{jobAddProgress}}, \code{\link{jobAdd}},
  \code{\link{jobRemove}}, \code{\link{jobRunScript}},
  \code{\link{jobSetProgress}}, \code{\link{jobSetState}}
}
\concept{jobs}
