% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{killTerminal}
\alias{killTerminal}
\title{Kill Terminal}
\usage{
killTerminal(id)
}
\arguments{
\item{id}{The terminal id. The \code{id}(s) are obtained from
\code{\link{getAllTerminals}()} or  \code{\link{createTerminal}()}.}
}
\description{
Kill processes and close a terminal.
}
\note{
The \code{kilTerminal} function was added in version 1.1.227 of RStudio.
}
