% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launcher-functions.R
\name{launcherAvailable}
\alias{launcherAvailable}
\title{Check if Workbench Launcher is Available}
\usage{
launcherAvailable()
}
\description{
Check if the Workbench launcher is available and configured to support
Workbench jobs; that is, jobs normally launched by the user through
the RStudio IDE's user interface.
}
\seealso{
Other job-launcher functionality: 
\code{\link{launcherConfig}()},
\code{\link{launcherContainer}()},
\code{\link{launcherControlJob}()},
\code{\link{launcherGetInfo}()},
\code{\link{launcherGetJobs}()},
\code{\link{launcherGetJob}()},
\code{\link{launcherHostMount}()},
\code{\link{launcherNfsMount}()},
\code{\link{launcherPlacementConstraint}()},
\code{\link{launcherResourceLimit}()},
\code{\link{launcherSubmitJob}()},
\code{\link{launcherSubmitR}()}
}
\concept{job-launcher functionality}
