% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launcher-functions.R
\name{launcherConfig}
\alias{launcherConfig}
\title{Define a Launcher Configuration}
\usage{
launcherConfig(name, value = NULL)
}
\arguments{
\item{name}{The name of the launcher configuration.}

\item{value}{The configuration value. Must either be an integer, float, or string.}
}
\description{
Define a launcher configuration, suitable for use with the \code{config} argument
to \code{\link[=launcherSubmitJob]{launcherSubmitJob()}}.
}
\seealso{
Other job submission: \code{\link{launcherContainer}},
  \code{\link{launcherHostMount}},
  \code{\link{launcherNfsMount}},
  \code{\link{launcherPlacementConstraint}},
  \code{\link{launcherResourceLimit}},
  \code{\link{launcherSubmitJob}}
}
\concept{job submission}
