% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launcher-functions.R
\name{launcherControlJob}
\alias{launcherControlJob}
\title{Interact with (Control) a Job}
\usage{
launcherControlJob(jobId, operation = c("Suspend", "Resume", "Stop",
  "Kill", "Cancel"))
}
\arguments{
\item{jobId}{The job ID.}

\item{operation}{The operation to execute.}
}
\description{
Interact with a job.
}
