% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launcher-functions.R
\name{launcherGetInfo}
\alias{launcherGetInfo}
\title{Retrieve Workbench Launcher Information}
\usage{
launcherGetInfo()
}
\description{
Retrieve information about the Workbench launcher, as well as the different clusters
that the launcher has been configured to use.
}
\seealso{
Other job-launcher functionality: 
\code{\link{launcherAvailable}()},
\code{\link{launcherConfig}()},
\code{\link{launcherContainer}()},
\code{\link{launcherControlJob}()},
\code{\link{launcherGetJobs}()},
\code{\link{launcherGetJob}()},
\code{\link{launcherHostMount}()},
\code{\link{launcherNfsMount}()},
\code{\link{launcherPlacementConstraint}()},
\code{\link{launcherResourceLimit}()},
\code{\link{launcherSubmitJob}()},
\code{\link{launcherSubmitR}()}
}
\concept{job-launcher functionality}
