% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launcher-functions.R
\name{launcherGetJob}
\alias{launcherGetJob}
\title{Retrieve Workbench Job Information}
\usage{
launcherGetJob(jobId)
}
\arguments{
\item{jobId}{The id of a Workbench job.}
}
\description{
Retrieve information on a Workbench job with id \code{jobId}.
}
