% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launcher-functions.R
\name{launcherGetJob}
\alias{launcherGetJob}
\title{Retrieve Job Information}
\usage{
launcherGetJob(jobId)
}
\arguments{
\item{jobId}{The id of a launcher job.}
}
\description{
Retrieve information on a job with id \code{jobId}.
}
