% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launcher-functions.R
\name{launcherPlacementConstraint}
\alias{launcherPlacementConstraint}
\title{Define a Launcher Placement Constraint}
\usage{
launcherPlacementConstraint(name, value = NULL)
}
\arguments{
\item{name}{The name of this placement constraint.}

\item{value}{The value of the constraint. A job will only be placed on a
requested node if the requested placement constraint is present.}
}
\description{
Define a launcher placement constraint, suitable for use with the
\code{placementConstraints} argument to
\code{\link[=launcherSubmitJob]{launcherSubmitJob()}}.
}
\seealso{
Other job submission: \code{\link{launcherConfig}()},
\code{\link{launcherContainer}()}, \code{\link{launcherHostMount}()},
\code{\link{launcherNfsMount}()}, \code{\link{launcherResourceLimit}()},
\code{\link{launcherSubmitJob}()}, \code{\link{launcherSubmitR}()}
}
