% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launcher-functions.R
\name{launcherResourceLimit}
\alias{launcherResourceLimit}
\title{Define a Launcher Resource Limit}
\usage{
launcherResourceLimit(type, value)
}
\arguments{
\item{type}{The resource limit type. Must be one of cpuCount, cpuFrequency,
cpuSet, cpuTime, memory, memorySwap. Different launcher plugins may support
different subsets of these resource limit types; please consult the plugin
documentation to learn which limits are supported.}

\item{value}{The formatted value of the requested limit.}
}
\description{
Define a launcher resource limit, suitable for use with the
\code{resourceLimits} argument to
\code{\link[=launcherSubmitJob]{launcherSubmitJob()}}.
}
\seealso{
Other job submission: \code{\link{launcherConfig}()},
\code{\link{launcherContainer}()}, \code{\link{launcherHostMount}()},
\code{\link{launcherNfsMount}()},
\code{\link{launcherPlacementConstraint}()},
\code{\link{launcherSubmitJob}()}, \code{\link{launcherSubmitR}()}
}
