% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefs.R
\name{readPreference}
\alias{readPreference}
\title{Read Preference}
\usage{
readPreference(name, default)
}
\arguments{
\item{name}{The name of the preference.}

\item{default}{The default value to use when the preference is not
available.}
}
\description{
Reads a user preference, useful to remember preferences across different R
sessions for the same user.
}
\details{
User preferences can have arbitrary names and values. You must write the
preference with \code{\link{writePreference}} before it can be read
(otherwise its default value will be returned).
}
\note{
The \code{readPreference} function was added in version 1.1.67 of
RStudio.
}
\seealso{
\code{\link{readRStudioPreference}}, which reads RStudio IDE
preferences.
}
