% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stubs.R
\name{replaceSelection}
\alias{replaceSelection}
\title{Replace Selection within a Document}
\usage{
replaceSelection(text, id = NULL)
}
\arguments{
\item{text}{A character vector, indicating what
text should be inserted. This will be coerced
into a length one character vector as with
\code{paste(text, collapse = "\\n")}.}

\item{id}{The document id. When \code{NULL} or blank,
the mutation will apply to the currently open, or last
focused, RStudio document. Use the \code{id} returned
from \code{\link{getActiveDocumentContext}()} to ensure
that the operation is applied on the intended document.}
}
\description{
Replaces the contents of the currently active selection
in the document with id \code{id} with \code{text}.
}
\note{
The \code{replaceSelection} function was added with version 0.99.796 of
RStudio.
}
\seealso{
Other Source Document Methods: \code{\link{getActiveDocumentContext}},
  \code{\link{replaceRanges}}
}

