% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stubs.R
\name{restartSession}
\alias{restartSession}
\title{Restart the R Session}
\usage{
restartSession(command = "")
}
\arguments{
\item{command}{A command (as a string) to be run after restarting.}
}
\description{
Restart the RStudio session.
}
\note{
The \code{restartSession} function was added in version 1.1.281 of
RStudio.
}
