% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stubs.R
\name{restartSession}
\alias{restartSession}
\title{Restart the R Session}
\usage{
restartSession(command = NULL)
}
\arguments{
\item{command}{The \R code to execute (as a string).}
}
\description{
Restart the \R session, and (optionally) run a command
after restarting \R.
}
