% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stubs.R
\name{savePlotAsImage}
\alias{savePlotAsImage}
\title{Save Active RStudio Plot as an Image}
\usage{
savePlotAsImage(file, format = c("png", "jpeg", "bmp", "tiff", "emf", "svg",
  "eps"), width, height)
}
\arguments{
\item{file}{Target filename}

\item{format}{Image format ("png", "jpeg", "bmp", "tiff", "emf", "svg", or "eps")}

\item{width}{Image width in pixels}

\item{height}{Image height in pixels}
}
\description{
Save the currnently active RStudio as an image file.
}
\note{
The \code{savePlotAsImage} function was introduced in RStudio 1.1.57
}

