% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stubs.R
\name{sendToConsole}
\alias{sendToConsole}
\title{Send Code to the Console}
\usage{
sendToConsole(code, echo = TRUE, execute = TRUE, focus = TRUE)
}
\arguments{
\item{code}{A character vector of code.}

\item{echo}{Whether to echo the code in the console before executing.}

\item{execute}{Execute the code sent to the console?}

\item{focus}{Focus the console?}
}
\description{
Send, and execute, within the RStudio console.
}
\note{
The \code{sendToConsole} function was added in version 0.99.853 of RStudio.
}

