% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{sendToTerminal}
\alias{sendToTerminal}
\title{Send Text to a Terminal}
\usage{
sendToTerminal(text, id = NULL)
}
\arguments{
\item{text}{Character vector containing text to be inserted.}

\item{id}{The terminal id. When \code{NULL} or blank,
the text will be inserted into the currently open, or last
focused, RStudio terminal. Use the \code{id} returned
from \code{\link{getActiveTerminalId}()} or
\code{\link{createTerminal}()} to ensure the text is inserted
to the intended terminal.}
}
\description{
Send text to specified terminal.
}
\note{
The \code{sendToTerminal} function was added in version 1.1.227 of RStudio.
}
\examples{
\dontrun{
rstudioapi::sendToTerminal("ls -l", id = "Terminal 1")
}

}
