% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{sendToTerminal}
\alias{sendToTerminal}
\title{Send Text to a Terminal}
\usage{
sendToTerminal(id, text)
}
\arguments{
\item{id}{The terminal id. The \code{id} is obtained from
\code{\link{getAllTerminals}()}, \code{\link{getVisibleTerminal}()},
or \code{\link{createTerminal}()}.}

\item{text}{Character vector containing text to be inserted.}
}
\description{
Send text to specified terminal.
}
\note{
The \code{sendToTerminal} function was added in version 1.1.233 of RStudio.
}
\examples{
\dontrun{
rstudioapi::sendToTerminal(id='My Terminal', 'ls -l\\n')
}

}
