% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialogs.R
\name{showDialog}
\alias{showDialog}
\title{Show Dialog Box}
\usage{
showDialog(title, message, url = "", timeout = 60)
}
\arguments{
\item{title}{The title to display in the dialog box.}

\item{message}{A character vector with the contents to display in the main
dialog area. Contents can contain the following HTML tags: "p", "em",
"strong", "b" and "i".}

\item{url}{An optional URL to display under the \code{message}.}

\item{timeout}{An optional timeout (in seconds). When set, if the user takes
longer than this timeout to provide a response, the request will be aborted.}
}
\description{
Shows a dialog box with a given title and contents.
}
\note{
The \code{showDialog} function was added in version 1.1.67 of RStudio.
}
\examples{
if (rstudioapi::isAvailable()) {
  rstudioapi::showDialog("Example Dialog", "This is an <b>example</b> dialog.")
}

}
