% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialogs.R
\name{showDialog}
\alias{showDialog}
\title{Show Dialog Box}
\usage{
showDialog(title, message, url = "", timeout = 10)
}
\arguments{
\item{title}{The title to display in the dialog box.}

\item{message}{A character vector with the contents to display in the main
dialog area. Contents can contain the following HTML tags: "p", "em",
"strong", "b" and "i".}

\item{url}{An optional url to display under the \code{message}.}

\item{timeout}{Timeout (in seconds) passed to \code{callFun()} if running as a job (see \code{?isJob}).}
}
\description{
Shows a dialog box with a given title and contents.
}
\details{
\preformatted{ showDialog("A dialog", "Showing <b>bold</b> text in the
message.") }
}
\note{
The \code{showDialog} function was added in version 1.1.67 of RStudio.
}
