% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialogs.R
\name{showPrompt}
\alias{showPrompt}
\title{Show Prompt Dialog Box}
\usage{
showPrompt(title, message, default = NULL)
}
\arguments{
\item{title}{The title to display in the dialog box.}

\item{message}{A character vector with the contents to display in the main
dialog area.}

\item{default}{An optional character vector that fills the prompt field with
a default value.}
}
\description{
Shows a dialog box with a prompt field.
}
\note{
The \code{showPrompt} function was added in version 1.1.67 of RStudio.
}
