% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialogs.R
\name{showQuestion}
\alias{showQuestion}
\title{Show Question Dialog Box}
\usage{
showQuestion(title, message, ok = NULL, cancel = NULL)
}
\arguments{
\item{title}{The title to display in the dialog box.}

\item{message}{A character vector with the contents to display in the main
dialog area.}

\item{ok}{And optional character vector that overrides the caption for the
OK button.}

\item{cancel}{An optional character vector that overrides the caption for
the Cancel button.}
}
\description{
Shows a dialog box asking a question.
}
\note{
The \code{showQuestion} function was added in version 1.1.67 of
RStudio.
}
