% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{terminalActivate}
\alias{terminalActivate}
\title{Activate Terminal}
\usage{
terminalActivate(id = NULL, show = TRUE)
}
\arguments{
\item{id}{The terminal id. The \code{id} is obtained from
\code{\link{terminalList}()}, \code{\link{terminalVisible}()},
or \code{\link{terminalCreate}()}. If NULL, the terminal tab will be
selected but no specific terminal will be chosen.}

\item{show}{If TRUE, bring the terminal to front in RStudio.
terminal buffer.}
}
\description{
Ensure terminal is running and optionally bring to front in RStudio.
}
\note{
The \code{terminalActivate} function was added in version 1.1.250 of RStudio.
}
\examples{
\dontrun{
rstudioapi::terminalActivate('Terminal 1', show=TRUE)
}

}
