% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{terminalBusy}
\alias{terminalBusy}
\title{Is Terminal Busy}
\usage{
terminalBusy(id)
}
\arguments{
\item{id}{The terminal id. The \code{id} is obtained from
\code{\link{terminalList}()}, \code{\link{terminalVisible}()},
or \code{\link{terminalCreate}()}.'}
}
\value{
a boolean
}
\description{
Are terminals reporting that they are busy?
}
\note{
The \code{terminalBusy} function was added in version 1.1.305 of RStudio.
}
