% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{terminalCreate}
\alias{terminalCreate}
\title{Create a Terminal}
\usage{
terminalCreate(caption = NULL, show = TRUE)
}
\arguments{
\item{caption}{The desired terminal caption. When \code{NULL} or blank,
the terminal caption will be chosen by the system.}

\item{show}{If \code{FALSE}, terminal won't be brought to front.}
}
\value{
The terminal identifier as a character vector (\code{NULL} if
  unable to create the terminal or the given terminal caption is already
  in use).
}
\description{
Create a new Terminal.
}
\note{
The \code{terminalCreate} function was added in version 1.1.305 of RStudio.
}
\examples{
\dontrun{
termId <- rstudioapi::terminalCreate('My Terminal')
}

}
