% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{terminalSend}
\alias{terminalSend}
\title{Send Text to a Terminal}
\usage{
terminalSend(id, text)
}
\arguments{
\item{id}{The terminal id. The \code{id} is obtained from
\code{\link{terminalList}()}, \code{\link{terminalVisible}()},
\code{\link{terminalCreate}()}, or \code{\link{terminalExecute}()}.}

\item{text}{Character vector containing text to be inserted.}
}
\description{
Send text to an existing terminal.
}
\note{
The \code{terminalSend} function was added in version 1.1.305 of RStudio.
}
\examples{
\dontrun{
termId <- rstudioapi::terminalCreate()
rstudioapi::terminalSend(termId, 'ls -l\\n')
}

}
