% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefs.R
\name{writePreference}
\alias{writePreference}
\title{Write Preference}
\usage{
writePreference(name, value)
}
\arguments{
\item{name}{The name of the preference.}

\item{value}{The value of the preference.}
}
\description{
Writes a user preference, useful to remember preferences across different R
sessions for the same user.
}
\note{
The \code{writePreference} function was added in version 1.1.67 of
RStudio.
}
\seealso{
\code{\link{writeRStudioPreference}}, which changes RStudio IDE
preferences.
}
